refactor export file
