+++ components/Layout/index.ts
@@ -1,2 +1,1 @@
 export { Layout } from './Layout'
-export * from './interface'
