+++ components/Menu/index.ts
@@ -1,3 +1,1 @@
 export { Menu } from './Menu'
-export * from './interface'
-export * from './styles'
