+++ containers/Home/Education/index.ts
@@ -1,3 +1,1 @@
 export { Education } from './Education'
-export * from './styles'
-export * from './interface'
