+++ containers/Home/Lib/index.ts
@@ -1,3 +1,2 @@
 export { Lib } from './Lib'
-export * from './styles'
-export * from './interface'
+export type { LibProps } from './interface'
