+++ containers/Home/index.ts
@@ -1,4 +1,2 @@
 export { HomePage } from './Home'
 export { SideBar } from './SideBar'
-export * from './styles'
-export * from './interface'
